# docs/swagger.yaml
basePath: /
definitions:
...
  entities.Student:
    properties:
      accountId:
        type: integer
      fio:
        type: string
      groupId:
        type: integer
      groupName:
        type: string
      id:
        type: integer
      phoneNumber:
        type: string
    type: object
...
  requests.CreateStudentRequest:
    properties:
      account_id:
        type: integer
      fio:
        type: string
      group_id:
        type: integer
      group_name:
        type: string
      phone_number:
        type: string
    type: object
...
  requests.UpdateStudentRequest:
    properties:
      account_id:
        type: integer
      fio:
        type: string
      group_id:
        type: integer
      group_name:
        type: string
      id:
        type: integer
      phone_number:
        type: string
    type: object
...
  usecases.ReadAllStudentsByGroupIdResponseDto:
    properties:
      students:
        items:
          $ref: '#/definitions/entities.Student'
        type: array
    type: object
...
  usecases.ReadStudentResponseDto:
    properties:
      student:
        $ref: '#/definitions/entities.Student'
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: Backend for KeenEye
  title: Backend for KeenEye
  version: 1.0.0
paths:
...
  /api/create-student:
    post:
      consumes:
      - application/json
      description: Create a new student (admin only)
      parameters:
      - description: Student info
        in: body
        name: student
        required: true
        schema:
          $ref: '#/definitions/requests.CreateStudentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.Student'
        "400":
          description: Invalid request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BasicAuth: []
      summary: Create student
      tags:
      - students
...
  /api/delete-student:
    delete:
      description: Delete student by ID (admin only)
      parameters:
      - description: Student ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid student ID
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BasicAuth: []
      summary: Delete student
      tags:
      - students
...
  /api/read-all-students:
    get:
      description: Returns list of all students (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.Student'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BasicAuth: []
      summary: Get all students
      tags:
      - students
  /api/read-all-students-by-group-id:
    get:
      description: Students of group (student sees own group, teacher sees own group,
        admin sees all)
      parameters:
      - description: Group ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecases.ReadAllStudentsByGroupIdResponseDto'
        "400":
          description: Invalid group ID
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BasicAuth: []
      summary: Get students by group ID
      tags:
      - students
...
  /api/read-student:
    get:
      description: Returns student by ID. Accessible by student (self), teacher of
        group, or admin
      parameters:
      - description: Student ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecases.ReadStudentResponseDto'
        "400":
          description: Invalid student ID
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BasicAuth: []
      summary: Get student by ID
      tags:
      - students
...
  /api/update-student:
    put:
      consumes:
      - application/json
      description: Update a student record.
      parameters:
      - description: Updated student info
        in: body
        name: student
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateStudentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Student'
        "400":
          description: Invalid request body
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Access forbidden
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BasicAuth: []
      summary: Update student
      tags:
      - students
...
securityDefinitions:
  BasicAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"